+++ src/music_player.h
@@ -9,9 +9,9 @@ class MusicPlayer {
     MusicPlayer(std::string filename);
     ~MusicPlayer();
     void pause(bool pause);
-    void setPosition(float time);
+    void setPosition(double time);
     bool isPlaying();
-    float getTime() const;
+    double getTime() const;
 
   private:
     static void callback(void *userData, Uint8 *stream, int len);
